interface/requirements.txt
streamlit>=1.0.0
pandas>=1.5.0
kafka-python>=2.0.0
plotly>=5.0.0
psycopg2-binary>=2.9.0
matplotlib>=3.5.0
